@@ -1,27 +1,24 @@
 {
   "name": "com.synplyworks.thenmore",
-  "version": "0.8.10",
+  "version": "0.8.9",
   "description": "Homey Timer addon",
   "main": "app.ts",
   "dependencies": {
     "homey-api": "^1.10.20"
   },
   "devDependencies": {
-    "@tsconfig/node12": "^1.0.11",
-    "@types/homey": "npm:homey-apps-sdk-v3-types@^0.3.4",
-    "@types/node": "^18.15.3",
-    "@types/node-fetch": "^2.6.2",
-    "@types/source-map-support": "^0.5.6",
-    "@types/ws": "^8.5.3",
+    "@types/homey": "npm:homey-apps-sdk-v3-types@^0.3.7",
+    "@types/node": "^22.7.3",
+    "@types/node-fetch": "^2.6.11",
+    "@types/source-map-support": "^0.5.10",
+    "@types/ws": "^8.5.12",
+    "@typescript-eslint/eslint-plugin": "^8.7.0",
+    "@typescript-eslint/parser": "^8.7.0",
+    "eslint": "^9.11.1",
+    "homey": "^3.7.4",
     "athom-api": "^3.8.3",
-    "babel-eslint": "^10.1.0",
-    "eslint": "^7.32.0",
-    "eslint-config-airbnb": "^19.0.4",
-    "eslint-config-airbnb-typescript": "^17.0.0",
-    "eslint-config-athom": "^3.1.1",
-    "prettier": "^2.7.1",
-    "ts-node": "^10.9.1",
-    "typescript": "^4.8.4"
+    "ts-node": "^10.9.2",
+    "typescript": "^5.6.2"
   },
   "scripts": {
     "lint": "eslint . --ext .ts",
